PYTHON-4636 Stop blocking the I/O Loop for socket reads
